# docs/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/auth/sign-up":{"post":{"tags":["auth-controller"],"summary":"Регистрация пользователя","description":"Создаёт нового пользователя и возвращает токен.","operationId":"signUp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpRequestDto"}}},"required":true},"responses":{"200":{"description":"Пользователь зарегистрирован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JwtAuthenticationResponseDto"}}}},"409":{"description":"Пользователь с таким email уже существует","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtAuthenticationResponseDto"}}}},"400":{"description":"Некорректные данные","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtAuthenticationResponseDto"}}}},"500":{"description":"Ошибка сервера","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtAuthenticationResponseDto"}}}}}}},"/auth/sign-in":{"post":{"tags":["auth-controller"],"summary":"Авторизация пользователя","description":"Авторизация пользователя и возвращает токен.","operationId":"signIn","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInRequestDto"}}},"required":true},"responses":{"404":{"description":"Пользователь не найден","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtAuthenticationResponseDto"}}}},"200":{"description":"Пользователь авторизован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JwtAuthenticationResponseDto"}}}},"400":{"description":"Некорректные данные","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtAuthenticationResponseDto"}}}}}}},"/api/cards":{"get":{"tags":["card-controller"],"summary":"Получить все карты","description":"Возвращает список всех карт, принадлежащих текущему авторизованному пользователю.","operationId":"getAllUserCards","responses":{"404":{"description":"Карты не найдены","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CardDto"}}}}},"200":{"description":"Список карт","content":{"application/json":{"schema":{"type":"string"}}}},"401":{"description":"Неавторизованный доступ","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CardDto"}}}}},"500":{"description":"Ошибка сервера","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CardDto"}}}}}}},"post":{"tags":["card-controller"],"summary":"Перевод средств","description":"Выполняет перевод средств с одной карты на другую.","operationId":"transfer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardTransferDto"}}},"required":true},"responses":{"404":{"description":"Карта не найдена","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardTransferDto"}}}},"200":{"description":"Перевод выполнен успешно","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardTransferDto"}}}},"400":{"description":"Некорректные данные перевода","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardTransferDto"}}}},"401":{"description":"Неавторизованный доступ","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardTransferDto"}}}},"500":{"description":"Ошибка сервера","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardTransferDto"}}}}}}},"/api/cards/{id}/block-request":{"post":{"tags":["card-controller"],"summary":"Запрос на блокировку карты","description":"Отправляет запрос администратору на блокировку карты.","operationId":"blockRequest","parameters":[{"name":"id","in":"path","description":"ID карты","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"404":{"description":"Карта не найдена","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRequestDto"}}}},"200":{"description":"Запрос на блокировку отправлен","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestDto"}}}},"401":{"description":"Неавторизованный доступ","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRequestDto"}}}},"500":{"description":"Ошибка сервера","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRequestDto"}}}}}}},"/api/admin/users":{"get":{"tags":["admin-controller"],"summary":"Получить всех пользователей","description":"Возвращает список всех пользователей.","operationId":"getAllUser","responses":{"200":{"description":"Список пользователей","content":{"application/json":{"schema":{"type":"string"}}}},"500":{"description":"Ошибка сервера","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}}},"post":{"tags":["admin-controller"],"summary":"Создать пользователя","description":"Создаёт нового пользователя.","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"Пользователь создан","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"400":{"description":"Некорректные данные","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"500":{"description":"Ошибка сервера","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/admin/cards/{userId}":{"post":{"tags":["admin-controller"],"summary":"Создать карту","description":"Создаёт новую карту для указанного пользователя.","operationId":"createCard","parameters":[{"name":"userId","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardDto"}}},"required":true},"responses":{"404":{"description":"Пользователь не найден","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDto"}}}},"200":{"description":"Карта успешно создана","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardDto"}}}},"400":{"description":"Некорректные данные","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDto"}}}},"500":{"description":"Ошибка сервера","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDto"}}}}}}},"/api/admin/users/{id}":{"get":{"tags":["admin-controller"],"summary":"Получить пользователя","description":"Возвращает пользователя по его ID.","operationId":"getUserById","parameters":[{"name":"id","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"404":{"description":"Пользователь не найден","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"200":{"description":"Пользователь найден","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"500":{"description":"Ошибка сервера","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"delete":{"tags":["admin-controller"],"summary":"Удалить пользователя","description":"Удаляет пользователя по его ID.","operationId":"deleteUser","parameters":[{"name":"id","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"204":{"description":"Пользователь удалён"},"404":{"description":"Пользователь не найден"},"500":{"description":"Ошибка сервера"}}},"patch":{"tags":["admin-controller"],"summary":"Обновить пользователя","description":"Обновляет информацию о пользователе.","operationId":"updateUser","parameters":[{"name":"id","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"404":{"description":"Пользователь не найден","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"200":{"description":"Пользователь обновлён","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"400":{"description":"Некорректные данные","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"500":{"description":"Ошибка сервера","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/admin/cards/{id}/block":{"patch":{"tags":["admin-controller"],"summary":"Заблокировать карту","description":"Заблокирует карту по её ID.","operationId":"blockCard","parameters":[{"name":"id","in":"path","description":"ID карты","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"Карта заблокирована","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardDto"}}}},"404":{"description":"Карта не найдена","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDto"}}}},"500":{"description":"Ошибка сервера","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDto"}}}}}}},"/api/admin/cards/{id}/active":{"patch":{"tags":["admin-controller"],"summary":"Активировать карту","description":"Активирует карту по её ID.","operationId":"activeCard","parameters":[{"name":"id","in":"path","description":"ID карты","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"Карта активирована","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardDto"}}}},"404":{"description":"Карта не найдена","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDto"}}}},"500":{"description":"Ошибка сервера","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDto"}}}}}}},"/api/cards/{id}/balance":{"get":{"tags":["card-controller"],"summary":"Получить баланс карты","description":"Возвращает текущий баланс указанной карты.","operationId":"getBalance","parameters":[{"name":"id","in":"path","description":"ID карты","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"Баланс карты","content":{"text/plain":{"schema":{"type":"string"}}}},"404":{"description":"Карта не найдена или не принадлежит пользователю","content":{"*/*":{"schema":{"type":"string"}}}},"401":{"description":"Неавторизованный доступ","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Ошибка сервера","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/admin/cards":{"get":{"tags":["admin-controller"],"summary":"Получить все карты","description":"Возвращает список всех карт.","operationId":"getAllCards","responses":{"200":{"description":"Список карт","content":{"application/json":{"schema":{"type":"string"}}}},"500":{"description":"Ошибка сервера","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CardDto"}}}}}}}},"/api/admin/cards/{id}":{"delete":{"tags":["admin-controller"],"summary":"Удалить карту","description":"Удаляет карту по её ID.","operationId":"deleteCard","parameters":[{"name":"id","in":"path","description":"ID карты","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"404":{"description":"Карта не найдена","content":{"*/*":{"schema":{"type":"integer","format":"int64"}}}},"500":{"description":"Ошибка сервера","content":{"*/*":{"schema":{"type":"integer","format":"int64"}}}},"200":{"description":"Карта удалена","content":{"application/json":{"schema":{"type":"number"}}}}}}}},"components":{"schemas":{"SignUpRequestDto":{"required":["email","password","username"],"type":"object","properties":{"username":{"maxLength":50,"minLength":5,"type":"string","description":"Имя пользователя","example":"user"},"email":{"maxLength":50,"minLength":5,"type":"string","description":"Адрес электронной почты","example":"user@example.com"},"password":{"maxLength":50,"minLength":5,"type":"string","description":"Пароль","example":"password"}},"description":"Данные для регистрации"},"JwtAuthenticationResponseDto":{"type":"object","properties":{"token":{"type":"string","description":"Токен доступа","example":"eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJhZG1pbiIsImV4cCI6MTYyMjUwNj..."}},"description":"Ответ с JWT токеном после успешной аутентификации"},"SignInRequestDto":{"required":["password","username"],"type":"object","properties":{"username":{"maxLength":50,"minLength":5,"type":"string","description":"Имя пользователя","example":"Jon"},"password":{"maxLength":50,"minLength":5,"type":"string","description":"Пароль","example":"my_1secret1_password"}},"description":"Данные для входа"},"CardTransferDto":{"required":["amount","cardNumberFrom","cardNumberTo"],"type":"object","properties":{"cardNumberTo":{"maxLength":16,"minLength":16,"pattern":"^[0-9]{16}$","type":"string","description":"Номер карты получателя","example":"1234567812345678"},"cardNumberFrom":{"maxLength":16,"minLength":16,"pattern":"^[0-9]{16}$","type":"string","description":"Номер карты отправителя","example":"8765432187654321"},"amount":{"type":"integer","description":"Сумма перевода","example":15000}},"description":"Перевод денежных средств между картами"},"UserRequestDto":{"required":["cardNumber"],"type":"object","properties":{"cardNumber":{"maxLength":16,"minLength":16,"type":"string","description":"Номер карты для блокировки","example":"1234567812345678"}},"description":"Запрос пользователя на блокировку карты"},"UserDto":{"required":["email","name","password","role"],"type":"object","properties":{"name":{"maxLength":50,"minLength":5,"type":"string","description":"Имя пользователя","example":"testuser"},"email":{"type":"string","description":"Адрес электронной почты","example":"testuser@example.com"},"password":{"maxLength":50,"minLength":5,"type":"string","description":"Пароль пользователя","example":"password"},"role":{"type":"string","description":"Роль пользователя","example":"ROLE_USER"}},"description":"Пользователь"},"CardDto":{"required":["balance","cardNumber","expiryDate","status"],"type":"object","properties":{"cardNumber":{"maxLength":16,"minLength":16,"pattern":"^[0-9]{16}$","type":"string","description":"Номер карты","example":"1234567812345678"},"status":{"type":"string","description":"Статус карты","example":"ACTIVE","enum":["ACTIVE","BLOCKED","EXPIRED"]},"balance":{"type":"integer","description":"Баланс карты","example":10000},"expiryDate":{"type":"object","properties":{"year":{"type":"integer","format":"int32"},"month":{"type":"string","enum":["JANUARY","FEBRUARY","MARCH","APRIL","MAY","JUNE","JULY","AUGUST","SEPTEMBER","OCTOBER","NOVEMBER","DECEMBER"]},"monthValue":{"type":"integer","format":"int32"},"leapYear":{"type":"boolean"}},"description":"Срок действия карты (ГГГГ-ММ)","example":"2025-12"}},"description":"Банковская карта"}}}}